TW-700: Refactor and format code for `ChatListController`
