+++ assets/l10n/intl_en.arb
@@ -2755,6 +2755,8 @@
     "editWorkIdentitiesDescriptions": "Edit your work identity settings such as Matrix ID, email or company name.",
     "copiedMatrixIdToClipboard": "Copied Matrix ID to clipboard.",
     "changeProfilePhoto": "Change profile photo",
+    "countPinChat": "PINNED CHATS ({countPinChat})",
+    "countAllChat": "ALL CHATS ({countAllChat})",
     "thisMessageHasBeenEncrypted": "This message has been encrypted",
     "roomCreationFailed": "Room creation failed",
     "errorGettingPdf": "Error getting PDF",
